+++ base/organise.py
@@ -186,7 +186,7 @@ class Organise:
             default = None,
             type = str,
             nargs = '+',
-            help = 'Prepare cto output for this NFDI4Culture feed and catalog ID, optionally disable license check by adding "no_license_check".'
+            help = 'Prepare cto output for this NFDI4Culture feed and catalog ID, optionally disable feed element license checks via no-license-check as a third argument'
         )
         available_args.add_argument(
             '-bu', '--ba_username',
@@ -246,12 +246,12 @@ class Organise:
             if 'beacon' in self.output or 'csv' in self.output or 'cto' in self.output or 'cto3' in self.output:
                 raise ValueError('Hydra Scraper called with extraction routine but no element markup.')
 
-        # Check given arguments in prepare
+        # Check prepare arguments
         if self.prepare != None:
-            if len(self.prepare) < 2 or (len(self.prepare) == 2 and 'no_license_check' in self.prepare[:2]):
-                raise ValueError('Hydra Scraper expects NFDI4Culture feed and catalog ID in prepare. "no_license_check" can optionally be added as the third argument.')
-            elif len(self.prepare) == 3 and self.prepare[2] != 'no_license_check':
-                raise ValueError('Hydra Scraper expects "no_license_check" as the optional third argument in prepare.')
+            if len(self.prepare) < 2 or (len(self.prepare) == 2 and 'no-license-check' in self.prepare[:2]):
+                raise ValueError('Hydra Scraper expects NFDI4Culture feed and catalog ID in prepare; no-license-check is an optional third argument.')
+            elif len(self.prepare) == 3 and self.prepare[2] != 'no-license-check':
+                raise ValueError('Hydra Scraper only allows no-license-check as the optional third prepare argument.')
 
         # Check Basic Auth data
         if self.ba_username != None and self.ba_password == None: